requirements.txt
# Core Framework
llama-index>=0.12.0
llama-index-vector-stores-qdrant>=0.4.0
llama-index-embeddings-openai>=0.3.0
llama-index-llms-openai>=0.3.0
llama-index-llms-anthropic>=0.6.0

# PDF Parsing
docling>=2.0.0

# Vector Store
qdrant-client>=1.12.0

# Data Processing
pandas>=2.0.0

# Environment
python-dotenv>=1.0.0
